Simplify IResampler contract
