+++ src/ImageSharp/Processing/Processors/Transforms/Automorphic/AutoOrientProcessor.cs
