+++ src/ImageSharp/Processing/Processors/Transforms/Automorphic/FlipProcessor.cs
