+++ src/ImageSharp/Processing/Processors/Transforms/Automorphic/FlipProcessor{TPixel}.cs
