+++ src/ImageSharp/Processing/Processors/Transforms/Automorphic/RotateProcessor.cs
