+++ src/ImageSharp/Processing/Processors/Transforms/Automorphic/RotateProcessor{TPixel}.cs
