+++ src/ImageSharp/Processing/Processors/Transforms/Resamplers/TriangleResampler.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System.Numerics;
 using System.Runtime.CompilerServices;
 using SixLabors.ImageSharp.PixelFormats;
 
@@ -36,48 +35,8 @@ namespace SixLabors.ImageSharp.Processing.Processors.Transforms
 
         /// <inheritdoc/>
         [MethodImpl(InliningOptions.ShortMethod)]
-        public void ApplyResizeTransform<TPixel>(
-            Configuration configuration,
-            Image<TPixel> source,
-            Image<TPixel> destination,
-            Rectangle sourceRectangle,
-            Rectangle destinationRectangle,
-            bool compand)
-            where TPixel : struct, IPixel<TPixel> => ResamplerExtensions.ApplyResizeTransform(
-                configuration,
-                in this,
-                source,
-                destination,
-                sourceRectangle,
-                destinationRectangle,
-                compand);
-
-        /// <inheritdoc/>
-        [MethodImpl(InliningOptions.ShortMethod)]
-        public void ApplyAffineTransform<TPixel>(
-            Configuration configuration,
-            ImageFrame<TPixel> source,
-            ImageFrame<TPixel> destination,
-            Matrix3x2 matrix)
-            where TPixel : struct, IPixel<TPixel> => ResamplerExtensions.ApplyAffineTransform(
-                configuration,
-                in this,
-                source,
-                destination,
-                matrix);
-
-        /// <inheritdoc/>
-        [MethodImpl(InliningOptions.ShortMethod)]
-        public void ApplyProjectiveTransform<TPixel>(
-            Configuration configuration,
-            ImageFrame<TPixel> source,
-            ImageFrame<TPixel> destination,
-            Matrix4x4 matrix)
-            where TPixel : struct, IPixel<TPixel> => ResamplerExtensions.ApplyProjectiveTransform(
-                configuration,
-                in this,
-                source,
-                destination,
-                matrix);
+        public void ApplyTransform<TPixel>(IResamplingImageProcessor<TPixel> processor)
+            where TPixel : struct, IPixel<TPixel>
+            => processor.ApplyTransform(in this);
     }
 }
