+++ tests/ImageSharp.Tests/Processing/Processors/Transforms/ResizeKernelMapTests.ReferenceKernelMap.cs
@@ -27,7 +27,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Transforms
             public ReferenceKernel GetKernel(int destinationIndex) => this.kernels[destinationIndex];
 
             public static ReferenceKernelMap Calculate<TResampler>(in TResampler sampler, int destinationSize, int sourceSize, bool normalize = true)
-                where TResampler : unmanaged, IResampler
+                where TResampler : struct, IResampler
             {
                 double ratio = (double)sourceSize / destinationSize;
                 double scale = ratio;
